// app/core/models/auth.user.ts
export class GintaaUser {
    userId?: string;
    username?: string;
    profileUrl?: string;
    phoneNo?: string;
    email?: string;
    dob?: string;
    providerData?: any;
    refreshtoken?: string;
    accessToken?: string;
    expirationTime?: string;
    deviceId?: string;
    isVerified: boolean;
    claim?: string;
    business?: any;
}